refactor: 명료한 구분을 위해 version 1 패키지 분리 및 이동
