+++ src/main/java/com/example/mohago_nocar/plan/application/v1/response/BusPathResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.mohago_nocar.plan.application.response;
+package com.example.mohago_nocar.plan.application.v1.response;
 
 import com.example.mohago_nocar.transit.domain.model.BusPath;
 import com.example.mohago_nocar.transit.domain.model.SubPath;
