+++ src/main/java/com/example/mohago_nocar/plan/application/v1/response/PlanTravelCourseResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.mohago_nocar.plan.application.response;
+package com.example.mohago_nocar.plan.application.v1.response;
 
 import lombok.Builder;
