+++ src/main/java/com/example/mohago_nocar/plan/domain/service/TravelCourseUseCaseV1.java
@@ -1,7 +1,7 @@
 package com.example.mohago_nocar.plan.domain.service;
 
-import com.example.mohago_nocar.plan.presentation.request.PlanTravelCourseRequestDto;
-import com.example.mohago_nocar.plan.application.response.PlanTravelCourseResponseDto;
+import com.example.mohago_nocar.plan.presentation.v1.PlanTravelCourseRequestDto;
+import com.example.mohago_nocar.plan.application.v1.response.PlanTravelCourseResponseDto;
 
 import java.util.concurrent.CompletableFuture;
